Remove handler store
