+++ engine/src/main/kotlin/world/gregs/voidps/engine/event/EventStore.kt
@@ -3,6 +3,11 @@ package world.gregs.voidps.engine.event
 import com.github.michaelbull.logging.InlineLogger
 import it.unimi.dsi.fastutil.objects.Object2ObjectOpenHashMap
 import kotlinx.coroutines.*
+import world.gregs.voidps.engine.entity.World
+import world.gregs.voidps.engine.entity.character.Character
+import world.gregs.voidps.engine.entity.character.npc.NPC
+import world.gregs.voidps.engine.entity.character.player.Player
+import world.gregs.voidps.engine.entity.item.floor.FloorItem
 import kotlin.coroutines.CoroutineContext
 import kotlin.reflect.KClass
 
@@ -85,5 +90,35 @@ class EventStore : CoroutineScope {
         }
         var events = EventStore()
             private set
+        val parents = Object2ObjectOpenHashMap(mapOf<KClass<out EventDispatcher>, List<KClass<out EventDispatcher>>>(
+            EventDispatcher::class to listOf(World::class, FloorItem::class, Character::class),
+            Character::class to listOf(Player::class, NPC::class)
+        ))
     }
 }
+
+@Suppress("UNCHECKED_CAST")
+inline fun <reified T : EventDispatcher, reified E : Event> addEvent(noinline condition: E.(T) -> Boolean = { true }, priority: Priority = Priority.MEDIUM, noinline block: suspend E.(T) -> Unit) {
+    val dispatcher = T::class
+    val event = E::class
+    val handler = EventHandler(event, condition as Event.(EventDispatcher) -> Boolean, priority, block as suspend Event.(EventDispatcher) -> Unit)
+    EventStore.events.add(dispatcher, event, handler)
+    for (parent in EventStore.parents[dispatcher] ?: return) {
+        EventStore.events.add(parent, event, handler)
+    }
+}
+
+inline fun <reified E : Event> on(noinline condition: E.(Player) -> Boolean = { true }, priority: Priority = Priority.MEDIUM, noinline block: suspend E.(Player) -> Unit) =
+    addEvent(condition, priority, block)
+
+inline fun <reified E : Event> onNPC(noinline condition: E.(NPC) -> Boolean = { true }, priority: Priority = Priority.MEDIUM, noinline block: suspend E.(NPC) -> Unit) =
+    addEvent(condition, priority, block)
+
+inline fun <reified E : Event> onCharacter(noinline condition: E.(Character) -> Boolean = { true }, priority: Priority = Priority.MEDIUM, noinline block: suspend E.(Character) -> Unit) =
+    addEvent(condition, priority, block)
+
+inline fun <reified E : Event> onFloorItem(noinline condition: E.(FloorItem) -> Boolean = { true }, priority: Priority = Priority.MEDIUM, noinline block: suspend E.(FloorItem) -> Unit) =
+    addEvent(condition, priority, block)
+
+inline fun <reified E : Event> onWorld(noinline condition: E.(World) -> Boolean = { true }, priority: Priority = Priority.MEDIUM, noinline block: suspend E.(World) -> Unit) =
+    addEvent(condition, priority, block)
